woman-global/backend/requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asarPy==1.0.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.112.2
Flask==3.1.0
greenlet==3.0.3
h11==0.14.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.5
jose==1.0.0
k2-connect==1.2.6
Mako==1.3.5
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.8.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.4
typing_extensions==4.12.2
urllib3==2.3.0
urlvalidator==0.1
uvicorn==0.30.6
Werkzeug==3.1.3
wheel==0.45.1
